requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.4
anyio==3.6.1
apache-airflow==2.0.1
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.1
apispec==3.3.2
appnope==0.1.2
argcomplete==1.12.2
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.10.0
blis==0.7.7
brotlipy==0.7.0
cached-property==1.5.2
catalogue==2.0.7
cattrs==1.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
coloredlogs==14.0
colorlog==4.7.2
commonmark==0.9.1
connexion==2.7.0
contextlib2==21.6.0
contextvars==2.4
croniter==0.3.37
cryptography==35.0.0
cssselect==1.1.0
cymem==2.0.6
dataclasses==0.8
datasets==2.1.0
decorator==5.1.1
defusedxml==0.6.0
dill==0.3.4
dnspython==1.16.0
docker-pycreds==0.4.0
docstring-parser==0.14.1
docutils==0.16
email-validator==1.1.2
en-core-web-md==3.3.0
en-core-web-sm==3.3.0
en-core-web-trf==3.3.0
entrypoints==0.3
fastapi==0.70.1
filelock==3.0.10
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fsspec==2022.1.0
gensim==3.8.0
gitdb==4.0.7
GitPython==3.1.18
graphviz==0.16
greenlet==1.1.2
grpcio==1.46.3
grpcio-health-checking==1.43.0
gunicorn==19.10.0
h11==0.13.0
httptools==0.3.0
huggingface-hub==0.6.0
humanfriendly==10.0
idna==3.3
idna-ssl==1.1.0
immutables==0.17
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
iso8601==0.1.14
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.8.1
langcodes==3.3.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==3.8.0
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.12.0
multidict==5.2.0
multiprocess==0.70.12.2
murmurhash==1.0.7
natsort==7.1.1
nest-asyncio==1.5.1
nltk==3.4.5
numpy==1.19.5
olefile==0.46
openapi-spec-validator==0.2.9
OTXv2==1.5.12
packaging==21.3
pandas==1.1.5
parso==0.8.3
pathtools==0.1.2
pathy==0.6.1
pendulum==2.1.2
pep562==1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.4.1
pip==21.3.1
preshed==3.0.6
prison==0.1.3
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.19.4
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==5.0.0
pycparser==2.21
pydantic==1.8.2
pyfunctional==1.4.3
Pygments==2.12.0
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.17.3
PySocks==1.7.1
python-daemon==2.2.4
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.2.1
readability-lxml==0.8.1
regex==2021.8.3
requests==2.28.1
requests-file==1.5.1
rich==9.2.0
Rx==1.6.1
sacremoses==0.0.43
scikit-learn==0.24.2
scipy==1.5.2
sentry-sdk==1.5.12
seqeval==1.2.2
setproctitle==1.2.2
setuptools==59.6.0
shortuuid==1.0.1
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.3.1
spacy==3.3.0
spacy-alignments==0.8.5
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-transformers==1.1.5
SQLAlchemy==1.3.23
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
srsly==2.4.3
starlette==0.19.1
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
thinc==8.0.15
threadpoolctl==3.1.0
tldextract==3.3.0
tokenizers==0.10.3
toposort==1.7
torch==1.10.2
torchvision==0.5.0
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
transformers==4.14.1
typer==0.4.1
typing==3.7.4.3
typing-compat==0.1.0
typing_extensions==4.1.1
unicodecsv==0.14.1
urllib3==1.26.8
uvicorn==0.16.0
uvloop==0.14.0
wandb==0.12.18
wasabi==0.9.1
watchdog==2.1.9
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==1.0.1
wheel==0.37.1
WTForms==2.3.3
xxhash==2.0.2
yarl==1.6.3
zipp==3.6.0
